+++ tests/testthat/test_inputs.R
@@ -24,7 +24,7 @@ test_that("test_inputs", {
   expect_equal(test_inputs("help"),
                c("all", "scalar", "numeric", "integer", "logical",
                  "character", "factor", "data.frame", "matrix", "array",
-                 "date", "raw", "list"))
+                 "date", "raw", "na", "list"))
   expect_equal(test_inputs("help"),
                test_inputs(c("scalar", "help")))
 })